+++ pkg/wharfapi/requestHandler.go
@@ -8,7 +8,7 @@ import (
 	"net/url"
 	"regexp"
 
-	log "github.com/sirupsen/logrus"
+	"github.com/iver-wharf/wharf-core/pkg/logger"
 )
 
 var redacted = "*REDACTED*"
@@ -30,13 +30,19 @@ func redactTokenInURL(urlStr string) string {
 
 	uri, err := url.Parse(urlStr)
 	if err != nil {
-		log.WithError(err).Warningln("Unable to redact token from URL: parse URL")
+		log.Warn().
+			WithError(err).
+			WithString("action", "parse URL").
+			Message("Unable to redact token from URL.")
 		return ""
 	}
 
 	params, err := url.ParseQuery(uri.RawQuery)
 	if err != nil {
-		log.WithError(err).Warningln("Unable to redact token from URL: parse query")
+		log.Warn().
+			WithError(err).
+			WithString("action", "parse query").
+			Message("Unable to redact token from URL.")
 		return ""
 	}
 
@@ -55,7 +61,11 @@ func redactTokenInURL(urlStr string) string {
 
 	sanitized, err := url.PathUnescape(newURLStr)
 	if err != nil {
-		log.WithError(err).WithField("new URL string", newURLStr).Warningln("Unable to redact token from URL: unescape path")
+		log.Warn().
+			WithString("action", "unescape path").
+			WithError(err).
+			WithString("newURL", newURLStr).
+			Message("Unable to redact token from URL.")
 		return newURLStr
 	}
 
@@ -63,15 +73,17 @@ func redactTokenInURL(urlStr string) string {
 }
 
 func doRequest(from string, method string, URLStr string, body []byte, authHeader string) (*io.ReadCloser, error) {
-	log.WithFields(log.Fields{
-		"method": method,
-		"body":   redactTokenInJSON(string(body)),
-		"url":    redactTokenInURL(URLStr),
-	}).Debugln(from)
+	var redactedURL = redactTokenInURL(URLStr)
+	var withRequestMeta = func(ev logger.Event) logger.Event {
+		return ev.
+			WithString("method", method).
+			WithString("url", redactedURL)
+	}
+	withRequestMeta(log.Debug()).Message(from)
 
 	req, err := http.NewRequest(method, URLStr, bytes.NewReader(body))
 	if err != nil {
-		log.WithError(err).Errorln("Unable to prepare http request")
+		log.Error().WithError(err).Message("Failed preparing HTTP request.")
 		return nil, err
 	}
 
@@ -82,14 +94,18 @@ func doRequest(from string, method string, URLStr string, body []byte, authHeade
 	client := &http.Client{}
 	response, err := client.Do(req)
 	if err != nil {
-		log.WithError(err).Errorln("Unable to send http request")
+		withRequestMeta(log.Error()).
+			WithError(err).
+			Message("Failed sending HTTP request.")
 		return nil, err
 	}
 
 	if isNonSuccessful(response.StatusCode) {
 		if response.StatusCode == http.StatusUnauthorized {
 			response.Body.Close()
-			log.WithField("response", response).Errorln("Unauthorized")
+			withRequestMeta(log.Error()).
+				WithInt("status", response.StatusCode).
+				Message("Unauthorized.")
 			realm := response.Header.Get("WWW-Authenticate")
 			return nil, &AuthError{realm}
 		}
@@ -101,6 +117,12 @@ func doRequest(from string, method string, URLStr string, body []byte, authHeade
 			}
 			return nil, newProblemError(prob)
 		}
+
+		withRequestMeta(log.Warn()).
+			WithInt("status", response.StatusCode).
+			WithString("Content-Type", response.Header.Get("Content-Type")).
+			Messagef("Non-2xx should have responded with a Content-Type of %q.", problemContentType)
+		return nil, fmt.Errorf("unexpected status code returned: %s", response.Status)
 	}
 
 	return &response.Body, nil